gui is working
multiple updates
